+++ soft/main/hal/imu.h
@@ -27,8 +27,8 @@ class imu
 
     esp_err_t init(void);
 
-    esp_err_t read_acc(float * acc_x, float * acc_y, float * acc_z);
+    esp_err_t read_acc(double * acc_x, double * acc_y, double * acc_z);
 
-    esp_err_t read_gyro(float * gyro_x, float * gyro_y, float * gyro_z);
+    esp_err_t read_gyro(double * gyro_x, double * gyro_y, double * gyro_z);
 
 };
